refactor(接口测试): 重构接口执行
