+++ backend/services/api-test/src/main/java/io/metersphere/api/controller/ApiExecuteResourceController.java
@@ -1,20 +1,15 @@
 package io.metersphere.api.controller;
 
+import io.metersphere.api.service.ApiExecuteResourceService;
 import io.metersphere.api.service.ApiExecuteService;
-import io.metersphere.api.service.definition.ApiReportService;
-import io.metersphere.api.service.scenario.ApiScenarioReportService;
-import io.metersphere.sdk.constants.ExecStatus;
+import io.metersphere.sdk.dto.api.task.GetRunScriptRequest;
+import io.metersphere.sdk.dto.api.task.GetRunScriptResult;
 import io.metersphere.sdk.file.FileRequest;
-import io.metersphere.sdk.util.LogUtils;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.annotation.Resource;
 import jakarta.servlet.http.HttpServletResponse;
-import org.apache.commons.lang3.StringUtils;
-import org.springframework.data.redis.core.StringRedisTemplate;
 import org.springframework.web.bind.annotation.*;
-
-import java.util.Optional;
 
 /**
  * @Author: jianxing
@@ -24,15 +19,10 @@ import java.util.Optional;
 @RequestMapping("/api/execute/resource")
 @Tag(name = "接口测试-执行-资源")
 public class ApiExecuteResourceController {
-
-    @Resource
-    private StringRedisTemplate stringRedisTemplate;
     @Resource
     private ApiExecuteService apiExecuteService;
     @Resource
-    private ApiReportService apiReportService;
-    @Resource
-    private ApiScenarioReportService apiScenarioReportService;
+    private ApiExecuteResourceService apiExecuteResourceService;
 
     /**
      * 获取执行脚本
@@ -44,13 +34,18 @@ public class ApiExecuteResourceController {
     @GetMapping("script")
     @Operation(summary = "获取执行脚本")
     public String getScript(@RequestParam("reportId") String reportId, @RequestParam("testId") String testId) {
-        String key = apiExecuteService.getScriptRedisKey(reportId, testId);
-        LogUtils.info("获取执行脚本: {}", key);
-        String script = stringRedisTemplate.opsForValue().get(key);
-        stringRedisTemplate.delete(key);
-        apiReportService.updateReportStatus(reportId, ExecStatus.RUNNING.name());
-        apiScenarioReportService.updateReportStatus(reportId, ExecStatus.RUNNING.name());
-        return Optional.ofNullable(script).orElse(StringUtils.EMPTY);
+        return apiExecuteResourceService.getRunScript(reportId, testId);
+    }
+
+    /**
+     * 获取执行脚本
+     *
+     * @return
+     */
+    @PostMapping("script")
+    @Operation(summary = "获取执行脚本")
+    public GetRunScriptResult getScript(@RequestBody GetRunScriptRequest request) {
+        return apiExecuteResourceService.getRunScript(request);
     }
 
     /**
